dir: move modal files to src/modals
