+++ src/modals/SignUpModal/components/SelectBox/SelectBox.tsx
