+++ src/modals/SignUpModal/components/SelectBox/SelectStyle.tsx
