+++ src/modals/SignUpModal/components/ShowStep/ShowStep.tsx
