+++ src/modals/SignUpModal/components/ShowStep/ShowStepStyle.tsx
