+++ src/modals/SignUpModal/components/Timer/Timer.tsx
