+++ src/modals/SignUpModal/layout/LayoutStyle.tsx
