+++ src/modals/SignUpModal/layout/SignUpLayout.tsx
@@ -10,7 +10,7 @@ import {
 } from './LayoutStyle';
 import { ShowStep } from '../components';
 import { EStep } from '../type';
-import { Button } from '../../Button';
+import { Button } from '../../../shared/Button';
 import { LOGO_SERO } from '../../../assets';
 import { BackIcon, XButtonIcon } from '../../../assets/icon';
 import {
@@ -20,7 +20,7 @@ import {
   $stepIndex,
 } from '../state';
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { Pressable } from '../../Pressable/Pressable';
+import { Pressable } from '../../../shared/Pressable/Pressable';
 import { useSignupModalFormSubmit } from '../container';
 import { useModal } from '../../../providers';
