+++ src/modals/SignUpModal/views/Nickname/NicknameView.tsx
