+++ src/modals/SignUpModal/views/PasswordVerification/PasswordVerificationView.tsx
