+++ src/modals/index.ts
@@ -1,0 +1,2 @@
+export * from './LoginModal/LoginModal';
+export * from './SignUpModal/SignUpModal';
